Update Today- & HistoryView

+++ app/src/main/java/com/pascalrieder/proteincounter/database/models/DayItem.kt
@@ -7,8 +7,8 @@ import androidx.room.PrimaryKey
 @Entity
 data class DayItem(
     @PrimaryKey(autoGenerate = true) val uid: Long = 0,
-    val dayId: Int,
-    val itemId: Int,
+    val dayId: Long,
+    val itemId: Long,
     @ColumnInfo(name = "amountInGram") val amountInGram: Float,
     @ColumnInfo(name = "isDeleted") var isDeleted: Boolean = false
 )
